requirements.txt
cloudpickle==3.1.1
jupyter==1.1.1
lightning==2.4.0
msgpack==1.1.0
pip==24.2
plotly==6.0.0
polars==1.12.0
pyarrow==17.0.0
seaborn==0.13.2
tensorboard==2.18.0
torchaudio
torchvision
numpy
scikit-learn